requirements.txt
Flask
paddlepaddle
paddleocr>=2.0.1
arabic-reshaper
python-bidi